wip matrices by reference

+++ includes/fdf.h
@@ -6,7 +6,7 @@
 /*   By: qsharoly <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/03 18:00:09 by qsharoly          #+#    #+#             */
-/*   Updated: 2021/02/06 22:32:57 by debby            ###   ########.fr       */
+/*   Updated: 2021/02/11 13:01:41 by debby            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,7 +71,7 @@ typedef struct	s_things
 int				fail(char *msg);
 float			ft_fmin(float a, float b);
 float			ft_fmax(float a, float b);
-char			*ft_mini_dtoa(float a);
+char			*ft_mini_dtoa(double a);
 t_state			init_state(void);
 int				init_cam(t_cam *cam, t_things *things);
 int				init_map(t_map *map, const char *filename);
